LayoutTests/svg/transforms/negative-scale-value.html

<!DOCTYPE HTML>
<script src='../../resources/js-test.js'></script>
<script>
description("SVG scale transform with negative scale values should be correctly converted to strings.");

var svg = document.createElementNS('http://www.w3.org/2000/svg', 'svg');
var g = document.createElementNS('http://www.w3.org/2000/svg', 'g');
var transform = svg.createSVGTransform();
transform.setScale(-2, -4);

var list = g.transform.baseVal;
list.appendItem(transform);

shouldBeEqualToString("g.getAttribute('transform')", "scale(-2 -4)");
</script>
